requirements.txt
pyquaternion
matplotlib
numpy
PyYAML